+++ src/test/kotlin/com/github/jensim/megamanipulator/actions/git/clone/RemoteCloneOperatorTest.kt
@@ -7,9 +7,7 @@ import com.github.jensim.megamanipulator.actions.git.localrepo.LocalRepoOperator
 import com.github.jensim.megamanipulator.actions.search.SearchResult
 import com.github.jensim.megamanipulator.actions.vcs.PullRequestWrapper
 import com.github.jensim.megamanipulator.settings.types.CloneType.HTTPS
-import com.github.jensim.megamanipulator.settings.types.MegaManipulatorSettings
 import com.github.jensim.megamanipulator.settings.types.codehost.CodeHostSettings
-import com.github.jensim.megamanipulator.settings.types.searchhost.SearchHostSettings
 import com.intellij.openapi.project.Project
 import io.mockk.coEvery
 import io.mockk.coVerify
@@ -39,6 +37,9 @@ class RemoteCloneOperatorTest {
 
     private val localRepoOperator: LocalRepoOperator = mockk()
     private val processOperator: ProcessOperator = mockk()
+    private val sparseConfigSetupOperator: SparseConfigSetupOperator = mockk {
+        coEvery { setupSparseDef(any(), any()) } returns emptyList()
+    }
 
     private val project: Project = mockk()
     private val codeHostSettings = mockk<CodeHostSettings> {
@@ -46,16 +47,11 @@ class RemoteCloneOperatorTest {
         every { cloneType } returns HTTPS
         every { baseUrl } returns "https://example"
     }
-    private val settings = mockk<MegaManipulatorSettings> {
-        every { resolveSettings(any(), any()) } returns (
-            mockk<SearchHostSettings>() to codeHostSettings
-            )
-    }
-
     private val remoteCloneOperator = RemoteCloneOperator(
         project = project,
         localRepoOperator = localRepoOperator,
         processOperator = processOperator,
+        sparseConfigSetupOperator = sparseConfigSetupOperator,
     )
 
     private val tempDirPath: Path = createTempDirectory(prefix = null, attributes = emptyArray())